refactored for registry and graph-node queries
