+++ .gitignore
@@ -1,3 +1,4 @@
+.envrc
 /target
 /Cargo.lock
 /.idea
