+++ examples/graphcast/src/graphql/client_registry.rs
@@ -1,0 +1,49 @@
+use anyhow::anyhow;
+use graphql_client::{GraphQLQuery, Response};
+use serde_derive::{Deserialize, Serialize};
+
+#[derive(GraphQLQuery, Serialize, Deserialize, Debug)]
+#[graphql(
+    schema_path = "src/graphql/schema_registry.graphql",
+    query_path = "src/graphql/query_registry.graphql",
+    response_derives = "Debug, Serialize, Deserialize"
+)]
+pub struct GraphAccount;
+
+pub async fn perform_operator_indexer_query(
+    registry_subgraph_endpoint: String,
+    variables: graph_account::Variables,
+) -> Result<reqwest::Response, reqwest::Error> {
+    let request_body = GraphAccount::build_query(variables);
+    let client = reqwest::Client::new();
+    client
+        .post(registry_subgraph_endpoint)
+        .json(&request_body)
+        .send()
+        .await?
+        .error_for_status()
+}
+
+// directly return indexer address
+pub async fn query_registry_indexer(
+    registry_subgraph_endpoint: String,
+    operator_address: String,
+) -> Result<String, anyhow::Error> {
+    let variables: graph_account::Variables = graph_account::Variables {
+        address: operator_address,
+    };
+    let queried_result =
+        perform_operator_indexer_query(registry_subgraph_endpoint, variables).await?;
+
+    let response_body: Response<graph_account::ResponseData> = queried_result.json().await?;
+
+    if let Some(data) = response_body.data {
+        Ok(data
+            .graph_account
+            .and_then(|x| x.gossip_operator_of)
+            .map(|x| x.id)
+            .unwrap())
+    } else {
+        Err(anyhow!("No response data from registry"))
+    }
+}
