+++ examples/graphcast/src/graphql/query_network.rs
