+++ examples/graphcast/src/graphql/query_proof_of_indexing.graphql
