+++ examples/graphcast/src/graphql/query_registry.graphql
