+++ examples/graphcast/src/graphql/schema_network.graphql
