+++ examples/graphcast/src/graphql/schema_proof_of_indexing.graphql
