+++ examples/graphcast/src/graphql/schema_registry.graphql
